Add better breakpoint support
* Added multiple breakpoints and a breakpoint sub debug menu. This
allows the user to add multiple breakpoints and remove them. The menu
looks a little rough for now but works fairly well.
* The debug option 'continue' has been renamed to 'step' because
continue is being reused for something else. 'step' just moves to the
next instruction. 'continue' turns of debug mode and the GB continues
execution until it hits a breakpoint.